refactor: renames and dir for interface
