+++ packages/salesforcedx-utils-vscode/src/telemetry/interfaces/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright (c) 2024, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+export * from './telemetryReporter';
