+++ packages/salesforcedx-utils-vscode/src/telemetry/interfaces/telemetryReporter.ts
