+++ packages/salesforcedx-utils-vscode/src/telemetry/reporters/appInsightsReporter.ts
@@ -10,7 +10,7 @@ import * as os from 'os';
 import * as path from 'path';
 import { Disposable, env, UIKind, version, workspace } from 'vscode';
 import { WorkspaceContextUtil } from '../../context/workspaceContextUtil';
-import { TelemetryReporter } from './telemetryReporter';
+import { TelemetryReporter } from '../interfaces';
 
 export class AppInsights extends Disposable implements TelemetryReporter {
   private appInsightsClient: appInsights.TelemetryClient | undefined;
